finish update email userv and admin

+++ blueprints/admin/views.py
@@ -13,6 +13,7 @@ class AdminViews:
         self.user_tools = UserTools()
         self.scan_tools = ScannerTools()
         self.sql_injection_tools = SQLInjectionTools()
+        self.admin_access_list = [1, 2, '1', '2']
         self.admin_bp = Blueprint('admin', __name__, template_folder='templates')
         self.admin_routes()
 
@@ -91,6 +92,9 @@ class AdminViews:
         @self.admin_bp.route('/activate-admin/<int:ida>', methods=['PUT'])
         def activate_admin(ida):
             if self.user_tools.check_user_in_session('admin'):
+                current_admin = session['admin']
+                if current_admin['id_admin'] not in self.admin_access_list:
+                    return {'status': 'failed', 'message': 'Vous n\'avez pas les droits pour effectuer cette action'}
                 status = self.admin_service.activate_admin(ida)
                 if status == 'success':
                     return {'status': 'success'}
@@ -100,12 +104,18 @@ class AdminViews:
         @self.admin_bp.route('/add-admin', methods=['POST'])
         def add_admin():
             if self.user_tools.check_user_in_session('admin'):
+                current_admin = session['admin']
+                if current_admin['id_admin'] not in self.admin_access_list:
+                    return {'status': 'failed', 'message': 'Vous n\'avez pas les droits pour effectuer cette action'}
                 data = request.get_json()
                 nom = data.get('nom_admin')
                 prenom = data.get('prenom_admin')
                 email = data.get('email_admin')
                 if self.sql_injection_tools.detect_sql_injection([nom, prenom, email]):
                     return {'status': 'error', 'message': 'Problème de sécurité détecté'}
+                status, admin = self.admin_service.find_admin_by_email(email)
+                if status == 'success' and admin is not None:
+                        return {'status': 'failed', 'message': 'Email déjà utilisé'}
                 status = self.admin_service.add_admin(nom, prenom, email, 2)
                 if status == 'success':
                     return {'status': 'success'}
@@ -115,8 +125,50 @@ class AdminViews:
         @self.admin_bp.route('/delete-admin/<int:ida>', methods=['DELETE'])
         def delete_admin(ida):
             if self.user_tools.check_user_in_session('admin'):
+                current_admin = session['admin']
+                if current_admin['id_admin'] not in self.admin_access_list:
+                    return {'status': 'failed', 'message': 'Vous n\'avez pas les droits pour effectuer cette action'}
                 status = self.admin_service.delete_admin(ida)
                 if status == 'success':
                     return {'status': 'success'}
                 return {'status': 'failed'}
             return {'status': 'failed'}
+
+        @self.admin_bp.route('/update-admin/<int:id_admin>', methods=['PUT'])
+        def update_admin(id_admin):
+            if self.user_tools.check_user_in_session('admin'):
+                current_admin = session['admin']
+                if current_admin['id_admin'] not in self.admin_access_list:
+                    return {'status': 'failed', 'message': 'Vous n\'avez pas les droits pour effectuer cette action'}
+                data = request.get_json()
+                email = data.get('email_admin')
+                if self.sql_injection_tools.detect_sql_injection([email]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
+                status, admin = self.admin_service.find_admin_by_email(email)
+                if status == 'success' and admin is not None:
+                        return {'status': 'failed', 'message': 'Email déjà utilisé'}
+                status, admin = self.admin_service.find_admin_by_id(id_admin)
+                if status == 'success':
+
+                    status = self.admin_service.update_admin(id_admin, email)
+                    if status == 'success':
+                        socketio.emit('close_admin_session', {'message': 'success', 'id_admin': id_admin})
+                        return {'status': 'success'}
+                    return {'status': 'failed', 'message': 'Erreur de mise à jour'}
+                return {'status': 'failed', 'message': 'Admin non trouvé'}
+            return {'status': 'failed', 'message': 'Erreur de session'}
+
+        @self.admin_bp.route('/recuperer-admin/<int:id_admin>', methods=['PUT'])
+        def recuperation_admin(id_admin):
+            if self.user_tools.check_user_in_session('admin'):
+                current_admin = session['admin']
+                if current_admin['id_admin'] not in self.admin_access_list:
+                    return {'status': 'failed', 'message': 'Vous n\'avez pas les droits pour effectuer cette action'}
+                if self.sql_injection_tools.detect_sql_injection([id_admin]):
+                    return {'status': 'error', 'message': 'Problème de sécurité détecté'}
+                status, mail = self.admin_service.envoyer_email_recuperation(id_admin)
+                if status == 'success':
+                    socketio.emit('close_admin_session', {'message': 'success', 'id_admin': id_admin})
+                    return {'status': 'success', 'email': mail}
+                return {'status': 'failed', 'message': 'Erreur lors de l\'envoi de l\'email'}
+            return {'status': 'failed', 'message': 'Erreur de session'}
